Sentiment Analysis with BERT & labelling the dataset
I have used Google BERT-based uncase model for marking the sentiments of Reesby's potential employees.  In this 1 marks lowest and 5 highest. The sentiment was marked based on the responses of candidates and recruiters during the interview. The final outcome is the dataset with sentiment scores from 1-5 next to each response.